fix(cudf): move shared type utilities below exec
Move CudfConfig and Velox-to-cuDF type mapping into a common lower-level target so CudfVector can validate schemas without depending on the execution module.
